// src/interfaces/usuario.ts
// Export uma interface chamada Usuario
// dentro dele mostra o que o user vai ter
export interface Usuario {
    key?: string; 
    nome: string;
    endereco: string;
    bairro: string;
    cep: string;
    cidade: string;
    uf: string;
  //  dataNascimento: string;
    email: string;
  //  telefone: undefined;
}